feat: refactor product imports; update paths for ProductList, ProductCard, and RelatedProducts components feat: add GenderFilter component; implement gender selection functionality for product filtering feat: create ProductList component; integrate ProductCard for displaying products in a grid layout
